+++ packages/library-base/src/commonMain/kotlin/io/realm/Configuration.kt
@@ -58,9 +58,11 @@ public data class LogConfiguration(
     public val loggers: List<RealmLogger>
 )
 
+/**
+ * Base configuration options shared between all realm configuration types.
+ */
 public interface Configuration {
-    // Public properties making up the RealmConfiguration
-    // TODO Add more elaborate KDoc for all of these
+
     /**
      * Path to the realm file.
      */
@@ -126,7 +128,7 @@ public interface Configuration {
     public abstract class SharedBuilder<T, S : SharedBuilder<T, S>>(
         public var schema: Set<KClass<out RealmObject>> = setOf()
     ) {
-        protected var path: String? = null
+        protected var directory: String? = null
         protected var name: String = Realm.DEFAULT_FILE_NAME
         protected var logLevel: LogLevel = LogLevel.WARN
         protected var removeSystemLogger: Boolean = false
@@ -147,19 +149,21 @@ public interface Configuration {
         public abstract fun build(): T
 
         /**
-         * Sets the absolute path of the realm file.
+         * Sets the path to the directory that contains the realm file. If the directory does not
+         * exists, it and all intermediate directories will be created.
          *
-         * If not set the realm will be stored at the default app storage location for the platform.
-         *
-         * @param path either the canonical absolute path or a relative path from the current directory ('./').
+         * If not set the realm will be stored at the default app storage location for the platform:
          * ```
-         * // For JVM platforms the current directory is obtained using
+         * // For Android the default directory is obtained using
+         * Context.getFilesDir()
+         *
+         * // For JVM platforms the default directory is obtained using
          *  System.getProperty("user.dir")
          *
-         * // For macOS the current directory is obtained using
+         * // For macOS the default directory is obtained using
          * platform.Foundation.NSFileManager.defaultManager.currentDirectoryPath
          *
-         * // For iOS the current directory is obtained using
+         * // For iOS the default directory is obtained using
          * NSFileManager.defaultManager.URLForDirectory(
          *      NSDocumentDirectory,
          *      NSUserDomainMask,
@@ -168,8 +172,10 @@ public interface Configuration {
          *      null
          * )
          * ```
+         *
+         * @param directoryPath either the canonical absolute path or a relative path from the current directory ('./').
          */
-        public fun path(path: String?): S = apply { this.path = path } as S
+        public fun directory(directoryPath: String?): S = apply { this.directory = directoryPath } as S
 
         /**
          * Sets the filename of the realm file.