+++ packages/library-base/src/commonMain/kotlin/io/realm/RealmConfiguration.kt
@@ -62,7 +62,7 @@ public interface RealmConfiguration : Configuration {
             }
             allLoggers.addAll(userLoggers)
             return RealmConfigurationImpl(
-                path,
+                directory,
                 name,
                 schema,
                 LogConfiguration(logLevel, allLoggers),