+++ packages/library-base/src/darwin/kotlin/io/realm/internal/platform/SystemUtils.kt
@@ -2,12 +2,17 @@ package io.realm.internal.platform
 
 import io.realm.log.LogLevel
 import io.realm.log.RealmLogger
+import kotlinx.cinterop.BooleanVar
+import kotlinx.cinterop.ObjCObjectVar
 import kotlinx.cinterop.ULongVar
 import kotlinx.cinterop.alloc
 import kotlinx.cinterop.memScoped
+import kotlinx.cinterop.pointed
 import kotlinx.cinterop.ptr
 import kotlinx.cinterop.value
+import platform.Foundation.NSError
 import platform.Foundation.NSProcessInfo
+import platform.Foundation.NSURL
 import platform.posix.pthread_threadid_np
 import kotlin.native.concurrent.ensureNeverFrozen
 import kotlin.native.concurrent.freeze
@@ -36,3 +41,31 @@ public actual val <T> T.isFrozen: Boolean
     get() = this.isFrozen
 
 public actual fun Any.ensureNeverFrozen(): Unit = this.ensureNeverFrozen()
+
+// Depend on filesystem API's to handle edge cases around creating paths.
+public actual fun prepareRealmFilePath(directoryPath: String, filename: String): String {
+    val dir = NSURL.fileURLWithPath(directoryPath, isDirectory = true)
+    val fm = platform.Foundation.NSFileManager.defaultManager
+    memScoped {
+        val isDir = alloc<BooleanVar>()
+        val exists = fm.fileExistsAtPath(dir.absoluteString!!, isDirectory = isDir.ptr)
+        if (!exists) {
+            val errorPtr = alloc<ObjCObjectVar<NSError?>>()
+            fm.createDirectoryAtURL(
+                url = dir,
+                withIntermediateDirectories = true,
+                attributes = mapOf<Any?, Any>(),
+                errorPtr.ptr
+            )
+            errorPtr.ptr.pointed.value?.let {
+                throw IllegalArgumentException("Directories for Realm file could not be created: $directoryPath. Underlying cause: $it")
+            }
+        }
+        if (!isDir.value) {
+            throw IllegalArgumentException("Provided directory is a file: $directoryPath")
+        }
+    }
+
+    return NSURL.fileURLWithPath(filename, dir).absoluteString
+        ?: throw IllegalArgumentException("Could not resolve path components: '${directoryPath}' and '$filename'.")
+}