+++ test/base/src/androidTest/kotlin/io/realm/test/shared/RealmInstantTests.kt
@@ -25,7 +25,8 @@ class RealmInstantTests {
     fun setup() {
         tmpDir = PlatformUtils.createTempDir()
         val configuration = RealmConfiguration.Builder(schema = setOf(Sample::class))
-            .path("$tmpDir/default.realm").build()
+            .directory(tmpDir)
+            .build()
         realm = Realm.open(configuration)
     }